requirements.txt
flask
requests
numpy
pandas
scipy
python-dotenv
